test: add netns pid param to port checks
All of our test in run in their own namespace, we do not want to check
the hostns. Pass along the netns pid and nsenter for the grep command to
make sure it will check in the correct netns.

Also change some of the arguments ordering. Also most importantly fix
wait_for_port, it does actually the inverse. It must wait until a port
is bound by the application.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>