fix aarch64 ncat flake
I am not not 100% sure because I was unable to reproduce even with
hack/get_ci_vm.sh. This flakes permanently in CI so there must be a
different in the environment which causes it.

However looking at the code we spawn the nc listener in the background
so there is no guarantee that it will already listen when we make the nc
connect call. To fix this we use the wait_for_port logic to ensure the
port is bound.
For now there is no sctp support, the /proc file format for it is
completely different and I didn't want to spend more time on it so I
just added a sleep and call it good enough.

Fixes #433

Signed-off-by: Paul Holzinger <pholzing@redhat.com>